Implement the event handler of the view in the presenter.
The cell now has its protocol to define its interface to display the speakers. The protocol is used in the cell, the presenter, the related protocols, and the tests. SpeakerDisplayData is now equatable so it can be tested if the data sent to be displayed is the expected one. A protocol has been defined for the interactor so the event handler can invoke it and the code of the interactor has been refactored.